requirements.txt
busylight-for-humans >= 0.7.6
paho-mqtt